+++ packages/extensions/modelerfour/src/modeler/modelerfour.ts
@@ -59,7 +59,7 @@ import {
   AnyObjectSchema,
 } from "@autorest/codemodel";
 import { Session, Channel } from "@autorest/extension-base";
-import { fail, minimum, pascalCase, KnownMediaType } from "@azure-tools/codegen";
+import { fail, minimum, pascalCase, KnownMediaType, shadowPosition } from "@azure-tools/codegen";
 import {
   Model as oai3,
   Dereferenced,
@@ -71,6 +71,7 @@ import {
   NumberFormat,
   MediaType,
   omitXDashProperties,
+  OpenAPI3Document,
 } from "@azure-tools/openapi";
 import * as OpenAPI from "@azure-tools/openapi";
 import { uniq, every } from "lodash";
@@ -147,7 +148,6 @@ export class ModelerFour {
   private apiVersionParameter!: "choice" | "constant" | undefined;
   private useModelNamespace!: boolean | undefined;
   private profileFilter!: Array<string>;
-  private apiVersionFilter!: Array<string>;
   private schemaCache = new ProcessingCache((schema: OpenAPI.Schema, name: string) =>
     this.processSchemaImpl(schema, name),
   );
@@ -158,8 +158,8 @@ export class ModelerFour {
   private ignoreHeaders: Set<string> = new Set();
   private specialHeaders: Set<string> = new Set();
 
-  constructor(protected session: Session<oai3>) {
-    this.input = session.model; //shadowPosition(session.model);
+  constructor(protected session: Session<OpenAPI3Document>) {
+    this.input = session.model; // shadowPosition(session.model);
 
     const i = this.input.info;
 
@@ -268,7 +268,6 @@ export class ModelerFour {
     }
 
     this.profileFilter = await this.session.getValue("profile", []);
-    this.apiVersionFilter = await this.session.getValue("api-version", []);
     this.ignoreHeaders = new Set(this.options["ignore-headers"] ?? []);
     this.specialHeaders = new Set(
       KnownSpecialHeaders.filter((x) => this.options["skip-special-headers"]?.map((x) => x.toLowerCase())?.includes(x)),
@@ -867,6 +866,7 @@ export class ModelerFour {
     this.schemaCache.set(schema, objectSchema);
     for (const [propertyName, propertyDeclaration] of Object.entries(schema.properties ?? {})) {
       this.use(<OpenAPI.Refable<OpenAPI.Schema>>propertyDeclaration, (pSchemaName, pSchema) => {
+        const property = this.resolve(propertyDeclaration);
         const pType = this.processSchema(pSchemaName || `type·for·${propertyName}`, pSchema);
         const prop = objectSchema.addProperty(
           new Property(
@@ -880,8 +880,8 @@ export class ModelerFour {
               required: schema.required ? schema.required.indexOf(propertyName) > -1 : undefined,
               serializedName: propertyName,
               isDiscriminator: discriminatorProperty === propertyName ? true : undefined,
-              extensions: this.interpret.getExtensionProperties(pSchema, propertyDeclaration),
-              clientDefaultValue: this.interpret.getClientDefault(pSchema, propertyDeclaration),
+              extensions: this.interpret.getExtensionProperties(property, propertyDeclaration),
+              clientDefaultValue: this.interpret.getClientDefault(property, propertyDeclaration),
             },
           ),
         );
@@ -1382,10 +1382,12 @@ export class ModelerFour {
   }
 
   getUniqueName(baseName: string): string {
-    let nameCount = this.uniqueNames[baseName];
-    if (typeof nameCount == "number") {
-      this.uniqueNames[baseName] = nameCount++;
-      return `${baseName}${nameCount}`;
+    const nameCount = this.uniqueNames[baseName];
+
+    if (typeof nameCount === "number") {
+      const newCount = nameCount + 1;
+      this.uniqueNames[baseName] = newCount;
+      return `${baseName}${newCount}`;
     } else {
       this.uniqueNames[baseName] = 0;
       return baseName;
@@ -1651,6 +1653,7 @@ export class ModelerFour {
         extensions: this.interpret.getExtensionProperties(httpOperation),
         apiVersions: this.interpret.getApiVersions(pathItem),
         deprecated: this.interpret.getDeprecation(httpOperation),
+        externalDocs: this.interpret.getExternalDocs(httpOperation),
         language: {
           default: {
             summary: httpOperation.summary,