+++ packages/libs/extension/src/exceptions.ts
@@ -1,6 +1,5 @@
+import { SystemRequirementError } from "@autorest/system-requirements";
 import { Extension } from "./extension";
-import { PackageManagerLogEntry } from "./package-manager";
-import { SystemRequirementError } from "./system-requirements";
 
 export class UnresolvedPackageException extends Error {
   constructor(packageId: string) {