+++ packages/extensions/modelerfour/CHANGELOG.md
@@ -1,6 +1,41 @@
 # Change Log - @autorest/modelerfour
 
-This log was last generated on Mon, 21 Mar 2022 15:38:03 GMT and should not be manually modified.
+This log was last generated on Fri, 27 May 2022 15:55:11 GMT and should not be manually modified.
+
+## 4.23.6
+Fri, 27 May 2022 15:55:11 GMT
+
+### Patches
+
+- Fix issue with having `application/json` and `application/x-www-form-urlencoded` content type for the request body. Json now takes precedence
+
+## 4.23.5
+Wed, 18 May 2022 02:34:13 GMT
+
+### Patches
+
+- Fix name bgeneration for contenttype and accept parameters only incrementing to 1
+
+## 4.23.4
+Mon, 09 May 2022 15:29:40 GMT
+
+### Patches
+
+- Fix `x-ms-client-flatten` defined at the root of the model would always flatten that model when referemced in properties
+
+## 4.23.3
+Tue, 03 May 2022 20:20:34 GMT
+
+### Patches
+
+- **Fix** issue when using unknown media types with a known one. Ignore those.
+
+## 4.23.2
+Thu, 07 Apr 2022 20:47:30 GMT
+
+### Patches
+
+- Modelerfour set correct known media type when `application/json` with `type: string`
 
 ## 4.23.1
 Mon, 21 Mar 2022 15:38:03 GMT